requirements.txt
huggingface_hub
llama-cpp-python
chromadb
sentence-transformers
gradio
datasets
python-dotenv
